source/client/requirements.txt
munch==2.5.0
pycryptodome==3.15.0
six==1.16.0
